requirements.new.txt
Django==4.2.9
djangorestframework==3.14.0
django-filter==23.5
django-redis==5.4.0
django-rq==2.9.0
django-tables2==2.7.0
django-otp==1.3.0
django-otp-yubikey==1.1.0
django-mptt==0.15.0
django-formtools==2.5.1
django_browser_reload==1.12.1
drf-yasg==1.21.7
gunicorn==21.2.0
Markdown==3.5.1
Pillow==10.2.0
PyYAML==6.0.1
bleach==6.1.0
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.24.0
netaddr==0.9.0
phonenumbers==8.13.27
psycopg2-binary==2.9.9
rq-scheduler==0.13.1
segno==1.6.1
swagger-spec-validator==3.0.3
typing-extensions==4.9.0
redis==5.0.1
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
python-dotenv==1.0.0
sentry-sdk==1.39.1
prometheus-client==0.19.0
celery==5.3.6
flower==2.0.1
